move validation hooks to their own files
